pyproject.toml
[project]
name = "wallboard"
version = "0.1.0"
description = "GNOME wallpaper dashboard generator (Pillow + HTML/Playwright renderers)"
authors = [
    { name = "Vince Veselosky", email = "vince@control-escape.com" }
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "pyyaml>=6.0.2",
  "pillow>=10.4.0",
  "psutil>=6.0.0",
  "requests>=2.32.3",
  "icalendar>=5.0.13",
  "python-dateutil>=2.9.0.post0",
  "caldav>=1.3.9",
  "playwright>=1.49.0",
]

[project.scripts]
wallboard = "wallboard.cli:main"

[build-system]
requires = ["uv_build>=0.9.14,<0.10.0"]
build-backend = "uv_build"
